Updates to allow user defined styles to be applied, or team styles for commits to source control systems.
Updated AStyle to latest version.
Removed excess batch files.
Provided a default style which is used when a user / team has no style in their documents/AStyle folder already.
The stylecode.bat is improved to now take a command line parameter. If this is not set it will use the styles for the current user. If you set it to anything else it will use that as the first part of the filename to use when applying styles.

e.g.

stylecode hendrik

will run the AStyle formatter against the source files using a style document in your documents/astyle folder called 'hendrik.astyle'. Calling it with no parameter will use the style 'documents/astyle/%username%' instead, which will be created if it does not exist.

Suggested use: just run it as is to apply your style to the code. Optionally, pass a paramter to the 'pre build step' that calls 'stylecode' to have it style it for a team prior to booking into Git. You can even add a new build config that passes the correct parameter when you use it for the build.